add links and stats for gnn
