+++ scripts/data_preprocessing/links_and_stats/pop_1pm_basecase_mean_mode_stats.csv
@@ -1,0 +1,7 @@
+mode,avg_total_travel_time,avg_total_routed_distance,avg_trip_count
+bike,1187.1343820571963,3681.6741650943295,390.80392156862746
+car,941.0464589203576,4835.211251303553,3750.470588235294
+car_passenger,423.1120030101593,4378.124316882189,938.0
+outside,0.7887997604933445,1057.635257825451,2756.235294117647
+pt,1602.5077192338636,5467.415175248752,4067.921568627451
+walk,1007.7128624499618,1209.8118585433865,3232.0
