+++ scripts/data_preprocessing/links_and_stats/pop_1pm_policy_in_1_2_3_4_mean_mode_stats.csv
@@ -1,0 +1,7 @@
+mode,avg_total_travel_time,avg_total_routed_distance,avg_trip_count
+bike,1200.5128107472415,3723.1538814734445,383.4761904761905
+car,1023.5912698764753,4851.836312079865,3750.095238095238
+car_passenger,424.79089247639354,4387.0963824646,938.0
+outside,0.7887582301518979,1057.576160307536,2756.714285714286
+pt,1601.8407213066766,5452.917726962467,4075.3809523809523
+walk,1006.1274809295222,1207.9076686782337,3233.7619047619046
