+++ scripts/data_preprocessing/links_and_stats/pop_1pm_policy_in_1_2_3_4_percentage_diff_mode_stats.csv
@@ -1,0 +1,7 @@
+mode,avg_total_travel_time,avg_total_routed_distance,avg_trip_count
+bike,1.1269514970042067,1.1266536504610043,-1.8750403176673818
+car,8.771597849783054,0.34383318354146764,-0.010008081152095743
+car_passenger,0.3967955185128433,0.20492943856833273,0.0
+outside,-0.0052650043175254575,-0.005587703083623908,0.017378472645684408
+pt,-0.04162213505628508,-0.2651609183058934,0.18337088431176657
+walk,-0.1573247280564796,-0.15739553647998414,0.05451438000942415
