+++ scripts/data_preprocessing/links_and_stats/pop_1pm_policy_in_5_6_7_mean_mode_stats.csv
@@ -1,0 +1,7 @@
+mode,avg_total_travel_time,avg_total_routed_distance,avg_trip_count
+bike,1187.7993142129155,3683.7359938344966,388.70588235294116
+car,1019.7093892296336,4840.653575765195,3749.725490196078
+car_passenger,423.7476483130566,4381.8079054316695,938.0
+outside,0.7888427988606591,1057.6790292790035,2755.9607843137255
+pt,1603.1344298562517,5464.151238849191,4070.5686274509803
+walk,1008.0896496429241,1210.2637867092874,3230.156862745098
