+++ scripts/data_preprocessing/links_and_stats/pop_1pm_policy_in_5_6_7_percentage_diff_mode_stats.csv
@@ -1,0 +1,7 @@
+mode,avg_total_travel_time,avg_total_routed_distance,avg_trip_count
+bike,0.05601153212048358,0.05600247734346428,-0.5368521398826013
+car,8.359091048440298,0.11255608449733497,-0.019866787958641437
+car_passenger,0.15023097864751694,0.08413622553558278,0.0
+outside,0.00545618412556884,0.004138615201104829,-0.009959592510377376
+pt,0.03910811878570963,-0.05969797966573513,0.06507153048239366
+walk,0.03739033280236675,0.03735524352067458,-0.05702776159969379
